Move more plotting dependent tests out of core
