+++ tests/core/test_reader.py
@@ -1,7 +1,6 @@
 import os
 import platform
 
-from PIL import Image, ImageSequence
 import numpy as np
 import pytest
 
@@ -21,28 +20,6 @@ pytestmark = pytest.mark.skipif(
     platform.system() == 'Darwin', reason='MacOS testing on Azure fails when downloading'
 )
 skip_windows = pytest.mark.skipif(os.name == 'nt', reason='Test fails on Windows')
-
-
-@pytest.fixture()
-def gif_file(tmpdir):
-    filename = str(tmpdir.join('sample.gif'))
-
-    pl = pyvista.Plotter(window_size=(300, 200))
-    pl.open_gif(filename, palettesize=16, fps=1)
-
-    mesh = pyvista.Sphere()
-    opacity = mesh.points[:, 0]
-    opacity -= opacity.min()
-    opacity /= opacity.max()
-    for color in ['red', 'blue', 'green']:
-        pl.clear()
-        pl.background_color = 'w'
-        pl.add_mesh(mesh, color=color, opacity=opacity)
-        pl.camera_position = 'xy'
-        pl.write_frame()
-
-    pl.close()
-    return filename
 
 
 def test_get_reader_fail():
@@ -930,27 +907,6 @@ def test_hdf_reader():
     assert mesh.n_points == 6724
     assert 'VEL' in mesh.point_data
     assert mesh.n_cells == 4800
-
-
-@pytest.mark.skipif(not HAS_IMAGEIO, reason="Requires imageio")
-def test_gif_reader(gif_file):
-    reader = pyvista.get_reader(gif_file)
-    assert isinstance(reader, pyvista.GIFReader)
-    assert reader.path == gif_file
-    reader.show_progress()
-
-    grid = reader.read()
-    assert grid.n_arrays == 3
-
-    img = Image.open(gif_file)
-    new_grid = pyvista.UniformGrid(dimensions=(img.size[0], img.size[1], 1))
-
-    # load each frame to the grid
-    for i, frame in enumerate(ImageSequence.Iterator(img)):
-        data = np.array(frame.convert('RGB').getdata(), dtype=np.uint8)
-        data_name = f'frame{i}'
-        new_grid.point_data.set_array(data, data_name)
-        assert np.allclose(grid[data_name], new_grid[data_name])
 
 
 def test_xdmf_reader():
